実装全体像変更
