+++ day5/演習3/tests/test_model.py
@@ -1,9 +1,9 @@
 import os
-import pytest
-import pandas as pd
-import numpy as np
-import pickle
 import time
+import pickle
+import pytest
+import numpy as np
+import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import accuracy_score
@@ -12,7 +12,7 @@ from sklearn.preprocessing import OneHotEncoder, StandardScaler
 from sklearn.compose import ColumnTransformer
 from sklearn.pipeline import Pipeline
 
-# テスト用データとモデルパスを定義
+# パス定義
 DATA_PATH = os.path.join(os.path.dirname(__file__), "../data/Titanic.csv")
 MODEL_DIR = os.path.join(os.path.dirname(__file__), "../models")
 MODEL_PATH = os.path.join(MODEL_DIR, "titanic_model.pkl")
@@ -28,7 +28,6 @@ def sample_data():
         df = titanic.data
         df["Survived"] = titanic.target
 
-        # 必要なカラムのみ選択
         df = df[
             ["Pclass", "Sex", "Age", "SibSp", "Parch", "Fare", "Embarked", "Survived"]
         ]
@@ -42,33 +41,23 @@ def sample_data():
 @pytest.fixture
 def preprocessor():
     """前処理パイプラインを定義"""
-    # 数値カラムと文字列カラムを定義
     numeric_features = ["Age", "Pclass", "SibSp", "Parch", "Fare"]
     categorical_features = ["Sex", "Embarked"]
 
-    # 数値特徴量の前処理（欠損値補完と標準化）
-    numeric_transformer = Pipeline(
-        steps=[
-            ("imputer", SimpleImputer(strategy="median")),
-            ("scaler", StandardScaler()),
-        ]
-    )
+    numeric_transformer = Pipeline([
+        ("imputer", SimpleImputer(strategy="median")),
+        ("scaler", StandardScaler()),
+    ])
 
-    # カテゴリカル特徴量の前処理（欠損値補完とOne-hotエンコーディング）
-    categorical_transformer = Pipeline(
-        steps=[
-            ("imputer", SimpleImputer(strategy="most_frequent")),
-            ("onehot", OneHotEncoder(handle_unknown="ignore")),
-        ]
-    )
+    categorical_transformer = Pipeline([
+        ("imputer", SimpleImputer(strategy="most_frequent")),
+        ("onehot", OneHotEncoder(handle_unknown="ignore")),
+    ])
 
-    # 前処理をまとめる
-    preprocessor = ColumnTransformer(
-        transformers=[
-            ("num", numeric_transformer, numeric_features),
-            ("cat", categorical_transformer, categorical_features),
-        ]
-    )
+    preprocessor = ColumnTransformer([
+        ("num", numeric_transformer, numeric_features),
+        ("cat", categorical_transformer, categorical_features),
+    ])
 
     return preprocessor
 
@@ -76,25 +65,19 @@ def preprocessor():
 @pytest.fixture
 def train_model(sample_data, preprocessor):
     """モデルの学習とテストデータの準備"""
-    # データの分割とラベル変換
     X = sample_data.drop("Survived", axis=1)
     y = sample_data["Survived"].astype(int)
     X_train, X_test, y_train, y_test = train_test_split(
         X, y, test_size=0.2, random_state=42
     )
 
-    # モデルパイプラインの作成
-    model = Pipeline(
-        steps=[
-            ("preprocessor", preprocessor),
-            ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
-        ]
-    )
+    model = Pipeline([
+        ("preprocessor", preprocessor),
+        ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
+    ])
 
-    # モデルの学習
     model.fit(X_train, y_train)
 
-    # モデルの保存
     os.makedirs(MODEL_DIR, exist_ok=True)
     with open(MODEL_PATH, "wb") as f:
         pickle.dump(model, f)
@@ -112,60 +95,63 @@ def test_model_exists():
 def test_model_accuracy(train_model):
     """モデルの精度を検証"""
     model, X_test, y_test = train_model
-
-    # 予測と精度計算
     y_pred = model.predict(X_test)
     accuracy = accuracy_score(y_test, y_pred)
-
-    # Titanicデータセットでは0.75以上の精度が一般的に良いとされる
     assert accuracy >= 0.75, f"モデルの精度が低すぎます: {accuracy}"
 
 
 def test_model_inference_time(train_model):
     """モデルの推論時間を検証"""
     model, X_test, _ = train_model
-
-    # 推論時間の計測
     start_time = time.time()
     model.predict(X_test)
     end_time = time.time()
-
     inference_time = end_time - start_time
-
-    # 推論時間が1秒未満であることを確認
-    assert inference_time < 1.0, f"推論時間が長すぎます: {inference_time}秒"
+    assert inference_time < 1.0, f"推論時間が長すぎます: {inference_time:.3f}秒"
 
 
 def test_model_reproducibility(sample_data, preprocessor):
     """モデルの再現性を検証"""
-    # データの分割
     X = sample_data.drop("Survived", axis=1)
     y = sample_data["Survived"].astype(int)
     X_train, X_test, y_train, y_test = train_test_split(
         X, y, test_size=0.2, random_state=42
     )
 
-    # 同じパラメータで２つのモデルを作成
-    model1 = Pipeline(
-        steps=[
-            ("preprocessor", preprocessor),
-            ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
-        ]
-    )
+    model1 = Pipeline([
+        ("preprocessor", preprocessor),
+        ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
+    ])
+    model2 = Pipeline([
+        ("preprocessor", preprocessor),
+        ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
+    ])
 
-    model2 = Pipeline(
-        steps=[
-            ("preprocessor", preprocessor),
-            ("classifier", RandomForestClassifier(n_estimators=100, random_state=42)),
-        ]
-    )
-
-    # 学習
     model1.fit(X_train, y_train)
     model2.fit(X_train, y_train)
 
-    # 同じ予測結果になることを確認
     predictions1 = model1.predict(X_test)
     predictions2 = model2.predict(X_test)
 
     assert np.array_equal(predictions1, predictions2), "モデルの予測結果に再現性がありません"
+
+
+def test_inference_time_and_accuracy():
+    """推論時間と精度の統合テスト"""
+    # モデルとテストデータの読み込み
+    with open(MODEL_PATH, "rb") as f:
+        model = pickle.load(f)
+    df = pd.read_csv(DATA_PATH)
+    X_test = df.drop("Survived", axis=1)
+    y_test = df["Survived"].astype(int)
+
+    # 推論時間の測定
+    start = time.time()
+    y_pred = model.predict(X_test)
+    duration = time.time() - start
+
+    assert duration < 1.0, f"Inference too slow: {duration:.3f} sec"
+
+    # 精度の確認（80%以上を期待）
+    acc = accuracy_score(y_test, y_pred)
+    assert acc >= 0.8, f"Low inference accuracy: {acc:.2f}"
